Adding pagination to the blog
This involved altering _config.yml to reflect the pagination rate
already in use (5 posts), as well as adjusting the code index.html
code to use jekyll's off the shelf pagination. The css changes are
to move everything to the bottom. The css isn't pretty, but it'll
do in a pinch.